docs: Added more styling info
As it was the reason why I forked the repo in the first place,
I have added the variables to "personalize" the commit messages.
This makes it easier for people who want "dynamic" commit messages to get  them.
Only a limited set of variables are available for now, more if requested.
For now it all has to do with time and such.
e.g This commit has been commited at ``full_time`` by ``username``.
For signoff and other stuff, I'll consider them in the future.

Signed-off-by: Gekko Wrld <108436933+gekkowrld@users.noreply.github.com>